cleaned react lib
