+++ packages/anon-aadhaar-react/src/index.ts
@@ -1,12 +1,10 @@
-import { LogInWithAnonAadhaar } from './components/LogInWithAnonAadhaar'
 import { useAnonAadhaar } from './hooks/useAnonAadhaar'
 import { AnonAadhaarProvider } from './provider/AnonAadhaarProvider'
 import { AnonAadhaarProof } from './components/AnonAadhaarProof'
-import { LogInWithAnonAadhaarV2 } from './components/LoginV2/LogInWithAnonAadhaarV2'
+import { LogInWithAnonAadhaar } from './components/LogInWithAnonAadhaar'
 
 export {
   LogInWithAnonAadhaar,
-  LogInWithAnonAadhaarV2,
   useAnonAadhaar,
   AnonAadhaarProvider,
   AnonAadhaarProof,
